requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pandas==2.2.2
pydantic==2.7.4
httpx==0.27.0